requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
firebase-admin==5.2.0
requests>=2.25.1
